backend/requirements.txt
Flask==2.3.0
flask-cors==4.0.0
ffmpeg-python==0.2.0
Pillow==10.0.0
python-dotenv==1.0.0
